Align OIDC authenticator class with the rest
The standard that we have for authenticators is that they
have an `Authenticator` class that is both a command-class
and a regular one. However, in OIDC we had a command-class
named `Authenticate`.
