+++ app/controllers/authenticate_controller.rb
@@ -87,7 +87,7 @@ class AuthenticateController < ApplicationController
   end
 
   def authenticate_oidc
-    authentication_token = Authentication::AuthnOidc::Authenticate.new.(
+    authentication_token = Authentication::AuthnOidc::Authenticator.new.(
       authenticator_input: oidc_authenticator_input
     )
     render json: authentication_token
